commit forward propagation tested

+++ src/DataLoader.cpp
@@ -1,22 +1,22 @@
 #include "DataLoader.h"
 
 
-DataLoader::DataLoader(int num_osc, int n, int tsw, int tst, float *o, float a_h, float a_k){
+DataLoader::DataLoader(int num_osc, int n, int tsw, int tst, double *o, double a_h, double a_k){
     Tsw = tsw; 
     Tst = tst;
     T = Tsw+Tst;
-    beta = (float) Tst/ (float) T;
+    beta = (double) Tst/ (double) T;
     offset = o;
     A_h = a_h;
     A_k = a_k;
     num_osc = num_osc;
     N = n;
-    heading = new float[2];
+    heading = new double[2];
     heading[0] = 0;
     heading[1] = 1;
 }
 
-void DataLoader::HipJoint(float *out, int off, float h){
+void DataLoader::HipJoint(double *out, int off, double h){
     int t;
     for(int j = T-off, k=0; j<N+T-off && k<N; j++,k++){
         t = j%T;
@@ -32,7 +32,7 @@ void DataLoader::HipJoint(float *out, int off, float h){
     }
 } 
 
-void DataLoader::KneeJoint(float *out, int off){
+void DataLoader::KneeJoint(double *out, int off){
     int t;
     for(int j = T-off, k = 0; j<N+T-off && k<N; j++, k++){
         t = j%T;
@@ -45,23 +45,23 @@ void DataLoader::KneeJoint(float *out, int off){
     }
 }
 
-void DataLoader::setHeading(float x, float y){
+void DataLoader::setHeading(double x, double y){
     heading[0] = x;
     heading[1] = y;
 }
 
-void DataLoader::getModelOutput(float **out){
+void DataLoader::getModelOutput(double **out){
     int off = T/4;
-    float steer_angle;
-    float temp;
+    double steer_angle;
+    double temp;
     for(int i = 0; i< 2*num_osc; i++){
         if(i<4){
             steer_angle = atan(heading[1]/heading[0]);
             if(steer_angle>0){
-                temp = (steer_angle)*180/(float)M_PI;
+                temp = (steer_angle)*180/(double)M_PI;
             }
             else{
-                temp = (steer_angle+M_PI)*180/(float)M_PI;
+                temp = (steer_angle+M_PI)*180/(double)M_PI;
             }
             HipJoint(out[i], i*off, temp);
         }
@@ -71,8 +71,8 @@ void DataLoader::getModelOutput(float **out){
     } 
 }
 
-void DataLoader::getModelInput(float *out){
-    out = new float[3];
+void DataLoader::getModelInput(double *out){
+    out = new double[3];
     out[0] = 2*6*A_h/T;
     out[1] = heading[0];
     out[2] = heading[1];
