+++ src/DataLoader.h
@@ -18,28 +18,28 @@ class DataLoader{
         int Tsw;
         int Tst;
         int T;
-        float beta;
-        float *offset;
-        float A_h;
-        float A_k;
+        double beta;
+        double *offset;
+        double A_h;
+        double A_k;
         int num_osc;
-        float *heading;
+        double *heading;
         int N;
-        void HipJoint(float *out, int off, float h);
-        void KneeJoint(float *out,  int off); 
+        void HipJoint(double *out, int off, double h);
+        void KneeJoint(double *out,  int off); 
     public:
-        DataLoader(int num_osc, int n, int tsw, int tst, float *o, float a_h, float a_k);
-        void getModelInput(float *out);
-        void setHeading(float x, float i);
-        void getModelOutput(float **out);
-        void setTsw(float tsw){ 
+        DataLoader(int num_osc, int n, int tsw, int tst, double *o, double a_h, double a_k);
+        void getModelInput(double *out);
+        void setHeading(double x, double i);
+        void getModelOutput(double **out);
+        void setTsw(double tsw){ 
             Tsw = tsw;
             T = Tsw+Tst;
-            beta = (float) Tst/ (float) T;
+            beta = (double) Tst/ (double) T;
         }
-        void setTst(float tst){
+        void setTst(double tst){
             Tst = tst;
             T = Tsw+Tst;
-            beta = (float) Tst/(float) T;
+            beta = (double) Tst/(double) T;
         }
 };
